Refactor Wallet Faucet to be its own component
Closes: #1167
